Refactor serializers into proper namespace
