+++ src/Serializer/RepresentationSerializerInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sentry;
+namespace Sentry\Serializer;
 
 
 interface RepresentationSerializerInterface
