+++ src/Serializer/SerializerInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sentry;
+namespace Sentry\Serializer;
 
 interface SerializerInterface
 {
